requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-chess==1.999
sqlalchemy==2.0.35
pydantic==2.9.0
httpx==0.27.0
